requirements.txt
aiogram==2.10.1
aiohttp==3.6.2
aioredis==1.3.1
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
async==0.6.2
async-timeout==3.0.1
asyncio==3.4.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.0
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
dnspython==2.0.0
entrypoints==0.3
fake-useragent==0.1.11
filelock==3.0.12
flake8==3.8.3
future==0.18.2
hiredis==1.1.0
idna==2.10
isort==4.3.21
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
jsonconfig-tool==1.5
keyring==21.4.0
kiwisolver==1.2.0
language==0.6
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
messages==0.5.0
more-itertools==8.4.0
multidict==4.7.6
packaging==20.4
parse==1.16.0
peewee==3.13.3
pefile==2019.4.18
pep8==1.7.1
pluggy==0.13.1
portend==2.6
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyee==7.0.2
pyinstaller-hooks-contrib==2020.7
pylint==2.5.3
pymongo==3.10.1
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
python-box==5.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-telegram-bot==12.8
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
requests==2.24.0
schedule==0.6.0
scraping==0.0.3
six==1.15.0
soupsieve==2.0.1
spark-parser==1.8.9
sqlparse==0.3.1
style==1.1.6
telebot==0.0.3
tempora==4.0.0
toml==0.10.1
tqdm==4.48.0
transliterate==1.10.2
typing-extensions==3.7.4.2
uncompyle6==3.7.4
update==0.0.1
urllib3==1.25.10
virtualenv==20.0.28
w3lib==1.22.0
wcwidth==0.2.5
websockets==8.1
wrapt==1.12.1
xdis==5.0.4
yarl==1.5.0
zc.lockfile==2.0
